fixed MSVC compile warnings

+++ kernels/geometry/trianglei.h
@@ -108,8 +108,8 @@ namespace embree
 	  geomID[i] = prim.geomID();
 	  primID[i] = prim.primID();
 	  v0[i] = (Vec3f*) mesh->vertexPtr(tri.v[0]); 
-	  v1[i] = (int*)   mesh->vertexPtr(tri.v[1])-(int*)v0[i]; 
-	  v2[i] = (int*)   mesh->vertexPtr(tri.v[2])-(int*)v0[i]; 
+	  v1[i] = int(size_t((int*)   mesh->vertexPtr(tri.v[1])-(int*)v0[i])); 
+	  v2[i] = int(size_t((int*)   mesh->vertexPtr(tri.v[2])-(int*)v0[i])); 
 	  prims++;
 	} else {
 	  assert(i);
@@ -143,8 +143,8 @@ namespace embree
 	  geomID[i] = prim->geomID();
 	  primID[i] = prim->primID();
 	  v0[i] = (Vec3f*) mesh->vertexPtr(tri.v[0]); 
-	  v1[i] = (int*)   mesh->vertexPtr(tri.v[1])-(int*)v0[i]; 
-	  v2[i] = (int*)   mesh->vertexPtr(tri.v[2])-(int*)v0[i]; 
+	  v1[i] = int(size_t((int*)   mesh->vertexPtr(tri.v[1])-(int*)v0[i])); 
+	  v2[i] = int(size_t((int*)   mesh->vertexPtr(tri.v[2])-(int*)v0[i])); 
 	  begin++;
 	} else {
 	  assert(i);
